Add comments and move the classes into new packages.
