+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/view/ui/ToolBoxPanel.java
@@ -17,8 +17,10 @@
  * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
  */
 
-package org.orbisgis.orbistoolbox.view;
+package org.orbisgis.orbistoolbox.view.ui;
 
+import org.orbisgis.orbistoolbox.view.ToolBox;
+import org.orbisgis.orbistoolbox.view.utils.TreeNodeWps;
 import org.orbisgis.sif.UIFactory;
 import org.orbisgis.sif.components.OpenFolderPanel;
 import org.orbisgis.sif.components.fstree.CustomTreeCellRenderer;
@@ -28,7 +30,6 @@ import org.orbisgis.sif.components.fstree.FileTreeModel;
 import javax.swing.*;
 import javax.swing.event.TreeSelectionEvent;
 import javax.swing.event.TreeSelectionListener;
-import javax.swing.tree.DefaultMutableTreeNode;
 import java.awt.*;
 import java.beans.EventHandler;
 import java.io.File;
@@ -36,15 +37,22 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
+ * Panel of the tool box containing the JTree of process.
+ *
  * @author Sylvain PALOMINOS
  **/
 
 public class ToolBoxPanel extends JPanel {
 
+    /** Reference to the toolbox.*/
     private ToolBox toolBox;
+    /** Root node of the JTree */
     private TreeNodeWps root;
+    /** Model of the Jtree */
     private FileTreeModel model;
+    /** JTree */
     private FileTree tree;
+    /** Panel containing the process information. */
     private ProcessInfoPanel processInfoPanel;
 
     public ToolBoxPanel(ToolBox toolBox){
@@ -71,6 +79,11 @@ public class ToolBoxPanel extends JPanel {
         tree.addTreeSelectionListener(EventHandler.create(TreeSelectionListener.class, this, "onNodeSelected", ""));
     }
 
+    /**
+     * Action done on selecting a node of the tree.
+     * If the process isn't a valid WPS process, change its icon.
+     * @param event
+     */
     public void onNodeSelected(TreeSelectionEvent event){
         TreeNodeWps selectedNode = (TreeNodeWps) ((FileTree)event.getSource()).getLastSelectedPathComponent();
         if(selectedNode != null) {
@@ -81,6 +94,13 @@ public class ToolBoxPanel extends JPanel {
         }
     }
 
+    /**
+     * Sets the information for the information panel.
+     * @param title
+     * @param abstrac
+     * @param inputList
+     * @param outputList
+     */
     public void setProcessInfo(String title, String abstrac, List<String> inputList, List<String> outputList){
         processInfoPanel.setTitle(title);
         processInfoPanel.setAbstrac(abstrac);
@@ -92,10 +112,11 @@ public class ToolBoxPanel extends JPanel {
         processInfoPanel.repaint();
     }
 
-    public void addSource(){
-
+    /**
+     * Adds a new local source for the toolBox.
+     */
+    public void addLocalSource(){
         File file = null;
-
         OpenFolderPanel openFolderPanel = new OpenFolderPanel("ToolBoxPanel.AddSource", "Add a source");
 
         //Wait the window answer and if the user validate set and run the export thread.
@@ -103,10 +124,10 @@ public class ToolBoxPanel extends JPanel {
             file = openFolderPanel.getSelectedFile();
         }
 
-        addSource(file);
+        addLocalSource(file);
     }
 
-    private TreeNodeWps addSource(File file){
+    private TreeNodeWps addLocalSource(File file){
         if(file == null) {
             return null;
         }
@@ -133,11 +154,14 @@ public class ToolBoxPanel extends JPanel {
         return source;
     }
 
+    /**
+     * Refreshes the selected source.
+     */
     public void refreshSource(){
         TreeNodeWps node = ((TreeNodeWps)tree.getLastSelectedPathComponent());
         if(!node.isRoot() && !node.isLeaf()) {
             root.remove(node);
-            addSource(node.getFilePath());
+            addLocalSource(node.getFilePath());
         }
     }
 
@@ -146,9 +170,6 @@ public class ToolBoxPanel extends JPanel {
         if (file.exists() && file.isDirectory()) {
             for (File f : file.listFiles()) {
                 if (f != null) {
-                    /*if (file.isDirectory()) {
-                        scriptList.addAll(getAllWpsScript(f));
-                    }*/
                     if (f.isFile() && f.getName().endsWith(".groovy")) {
                         scriptList.add(f);
                     }
